chatssd/requirements.txt
langchain
pydantic
dotenv
africastalking
flask
phonenumbers
pycountry
langchain_google_genai
google-adk
